requirements.txt
fastapi==0.115.0
uvicorn==0.24.0
python-dotenv==1.0.0
firebase-admin==6.2.0
pydantic==2.10.6
pydantic-settings==2.7.0